Test to compare mean energy/enstrophy and area under spectrum
