+++ py2d/aposteriori_analysis.py
@@ -22,39 +22,4 @@ def eddyTurnoverTime_2DFHIT(Omega):
     return eddyTurnoverTime
 
 
-def energy_2DFHIT(Psi, Omega, spectral = False):
-    '''Calculates energy as the mean of 0.5 * Psi * Omega.
-
-    Args:
-        Psi (np.ndarray): The stream function matrix.
-        Omega (np.ndarray): The vorticity matrix.
-        spectral (bool): Whether to perform inverse Fast Fourier Transform on Psi and Omega. Default is False.
-
-    Returns:
-        float: The calculated energy.
-    '''
-    if spectral:
-        Psi = np.fft.ifft2(Psi).real
-        Omega = np.fft.ifft2(Omega).real
-        
-    energy = np.mean(0.5 * Psi * Omega)
-    return energy
-
-
-def enstrophy_2DFHIT(Omega, spectral = False):
-    '''Calculates enstrophy as the mean of 0.5 * Omega * Omega.
-
-    Args:
-        Omega (np.ndarray): The vorticity matrix.
-        spectral (bool): Whether to perform inverse Fast Fourier Transform on Omega. Default is False.
-
-    Returns:
-        float: The calculated enstrophy.
-    '''
-    if spectral:
-        Omega = np.fft.ifft2(Omega).real
-        
-    enstrophy = np.mean(0.5 * Omega * Omega)
-    return enstrophy
-
